+++ devtools/replidev/src/command/mod.rs
@@ -1,6 +1,7 @@
 pub mod cargo;
 pub mod certs;
 pub mod conf;
+pub mod curl;
 pub mod deps;
 pub mod play;
 pub mod release;